Force float dtype in np.arange of scan methods

+++ motor.py
@@ -122,7 +122,7 @@ class Motor():
         spectra : ndarray
         """
         spectra = []
-        for position in np.arange(start, stop, step):
+        for position in np.arange(start, stop, step, dtype=np.float32):
             self.amove(position)
             if det is not None:
                 assert acq_time is not None
